+++ smartcontract/programs/doublezero-serviceability/src/processors/contributor/delete.rs
@@ -65,7 +65,9 @@ pub fn process_delete_contributor(
     }
 
     let contributor = Contributor::try_from(contributor_account)?;
-    if contributor.status != ContributorStatus::Activated {
+    if contributor.status != ContributorStatus::Activated
+        && contributor.status != ContributorStatus::Suspended
+    {
         return Err(DoubleZeroError::InvalidStatus.into());
     }
     if contributor.reference_count > 0 {