+++ smartcontract/programs/doublezero-serviceability/src/processors/device/delete.rs
@@ -76,6 +76,12 @@ pub fn process_delete_device(
 
     let mut device: Device = Device::try_from(device_account)?;
 
+    if device.status != DeviceStatus::Activated
+        && device.status != DeviceStatus::Suspended
+    {
+        return Err(DoubleZeroError::InvalidStatus.into());
+    }
+
     if device.reference_count > 0 {
         return Err(DoubleZeroError::ReferenceCountNotZero.into());
     }