+++ smartcontract/programs/doublezero-serviceability/src/processors/exchange/delete.rs
@@ -1,7 +1,7 @@
 use core::fmt;
 
 use crate::{
-    error::DoubleZeroError, globalstate::globalstate_get, helper::*, state::exchange::Exchange,
+    error::DoubleZeroError, globalstate::globalstate_get, helper::*, state::exchange::*,
 };
 use borsh::BorshSerialize;
 use borsh_incremental::BorshDeserializeIncremental;
@@ -63,6 +63,12 @@ pub fn process_delete_exchange(
     }
 
     let exchange = Exchange::try_from(exchange_account)?;
+
+    if exchange.status != ExchangeStatus::Activated
+        && exchange.status != ExchangeStatus::Suspended
+    {
+        return Err(DoubleZeroError::InvalidStatus.into());
+    }
 
     if exchange.reference_count > 0 {
         return Err(DoubleZeroError::ReferenceCountNotZero.into());